requirements.txt
Django==5.2.12
djangorestframework==3.16.1
coverage>=7.6,<8
drf-spectacular>=0.27,<1
django-tenants>=3.7,<4
psycopg[binary]>=3.2,<4
